Use bytes.Split instead of bytes.Fields.
Looking at the pprof statistics, bytes.Fields and
the functions called from it eat about 35% of the
cpu time. bytes.Split should do the same in the
regulare cases with only 3% cpu usage.
If the resulting array length is != 3, we use
bytes.Fields to give it a try again.

Another bytes.Fields call is in the metrics20 code,
replacing that will save us another few % cpu time.